docker/requirements.txt
requests
pandas
mysql-connector-python
sqlalchemy
pymysql